phase-3/backend/requirements.txt
fastapi
openai-agents
mcp
sqlmodel
psycopg2-binary
uvicorn
python-dotenv
python-multipart
sse-starlette
jsonschema
pydantic-settings
pydantic>=2.7.0
better-auth